+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Enable metadata aggregator to append Additional Attribute(s) into CMR.JSON as its own root key
 - **PODAAC-5089**
   - For each PNG file for granule, append metadata and mimetype for `cmr.json` file
+- **PODAAC-5053**
+  - Extract MGRS_TILE_ID metadata from iso.xml. Used for OPERA processing
 ### Deprecated
 ### Removed
 ### Fixed